requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.13.0
google-oauth2-tool==0.0.3
google-auth-oauthlib==1.1.0
Pillow==10.2.0
